修改了身高体重

+++ lib/pages/activity/entity/activity_questionnaire_entity.dart
@@ -104,8 +104,8 @@ class ActivityIllnessCaseEntity {
   int fillingDate;
   int birthDay;
   String nation;
-  int height;
-  int weight;
+  double height;
+  double weight;
   String status;
   int schedule;
   int completeTime;
